# Conflicts:
#	src/Main.java

+++ src/Main.java
@@ -1,10 +1,10 @@
 public class Main {
     public static void main(String[] args) {
 
-        int[] array = new int[]{1, 2, 3, 4};
+        long[] array = new long[]{1, 2, 3, 4};
 
-        int max = new SalesManager(array).max();
-        int mean = new SalesManager(array).mean();
+        long max = new SalesManager(array).max();
+        long mean = new SalesManager(array).mean();
         System.out.println(max);
         System.out.println(mean);
